datarun.py puts .npz files into ReSCAL output directory
created functions in ReSCAL to deal with timing for performance data
on writing and compressing .csp and ALTI files.
added a flag to rescal to determine if that performance data is
written to stdout.
example_pyrescal is now example_pyrescal.py.
in genesis.c main now returns 0 on success.
removed some whitespace in rescal_utilities.py